requirements.txt
requests
python-dotenv
configparser
python-oauth2
plurk-oauth